Add support for IPCA stock type in FixedIncome service
This update adds support for the "IPCA" stock type in the FixedIncome service. Previously, the service only handled "PRÉ" and other types, causing errors whenever an IPCA type was processed. Now, the function checks for the stock type and, if it's IPCA, it applies a different calculation using specific 'tx_quotient' and 'date_mask'. Additionally, the method 'get_tax_price' is called to get ipc tax value. This upgrade aims to increase the flexibility and accuracy of the FixedIncome service by including all types of stocks.