Corrected the computation of 'pvp' attribute based on investment type
This commit modifies the HttpRepository.py and AttStocks.py files to correct the computation of 'pvp' attribute based on the investment type. The 'get_info' function has been added to fetch the required information based on the investment type from the website. If the investment type is less than or equal to 4, 'pvp' value will be fetched using 'get_info', else it will be computed from 'find_value'. The idea is to enhance the robustness and the accuracy of the 'pvp' value in various investment scenarios.