Update default date attributes in InvestmentHandler
This commit modifies the default 'data_fim' and 'data_ini' values in the InvestmentHandler class. 'data_fim' is now set to next day date and 'data_ini' is moved a month earlier in absence of these fields in arguments. The modification enhances data comprehensiveness and accuracy in the investment scenarios when initial or final dates aren't provided explicitly.