Expand result set inclusion in InvestmentHandler
The commit updates the InvestmentHandler class to include 'carteira' type in result set when the type is not 'carteira'. Added code to reverse the sorted result if 'sentido' is in arguments and default to true otherwise. Implemented sorting by 'sorted_by' parameter, or default to 'ganho_mensalizado_medio'. These changes allow for more versatility in data output by accommodating more types of results and offering different sorting options.