Update 're_add_resumes_period' for configurable date ranges
Altered the 're_add_resumes_period' to accept 'args' parameter for flexibility in providing the date ranges. This enables the users to specify the range of dates for re-calculation of investment summaries. A corresponding change has been made in 'app.py' to pass 'args' from the user's request to the modified function. This enhanced functionality gives users more control and improves the overall user experience.