Refactor InvestmentHandler and app.py to improve performance
Refactored InvestmentHandler's method to handle 'tipo' input in a more comprehensive manner. If 'tipo' equals to 'all', it will return both 'tipo' and 'ativo'; else it returns a list of 'tipo'. This change offers more variety in selecting data from the database. In 'app.py', multithreading is implemented to run re_add_resumes_period function. This change is made to improve app performance.