Commit message: Add default 'tipo' in InvestmentHandler, refactor function in app.py
Checked for the presence of 'tipo' in InvestmentHandler.py, and if not present, assigned a default value of 'all'. This change was made to avoid KeyErrors when 'tipo' is absent.

In app.py, refactored the function called in the threading.Thread from `att_stocks.re_add_resumes_period` to `re_add_resume_invest_period_tr` for better readability and to avoid naming confusion.