Correct handling of absent 'tipo' in InvestmentHandler
This commit introduces a change in the way 'tipo' is managed in InvestmentHandler.py. Previously, the absence of 'tipo' from the args dictionary could lead to KeyErrors. With this fix, if 'tipo' is not present, it is assigned a default value of 'all'. This makes the handling of 'tipo' more robust.
For better readability and to avoid naming confusion, the function in app.py called in the threading.Thread has also been refactored from `att_stocks.re_add_resumes_period` to `re_add_resume_invest_period_tr`.