Adjust date calculation in InvestmentHandler
This update corrects the date calculation used in the "re_add_resumes_period" function in the InvestmentHandler.py file. Previously, the calculation was incrementing the current date by one day. However, this caused an off by one error, as the calculation actually intended to decrement the current date by one day. Additionally, the dates for 'data_ini' and 'data_fim' are now set to the end of the day to match the system's timestamp, ensuring data consistency across the application.