Update Jenkinsfile to securely update remote service
The 'Jenkinsfile' was modified to securely update the remote service based on the environment branch. Previously, the service was updated without credentials, which poses a security risk. With the added step, credentials are now incorporated for SSH remote command execution thus enhancing build security. The URL of 'HttpRepository.py' was also corrected to improve system response.