Refactor conditional checks and increase timeout in InvestmentHandler
The refactoring concentrates on improving the clarity and arrangement of conditional checks specifically related to investment type and att_price_ in InvestmentHandler.py. Additionally, the timeout in the get_values_by_ticker function has been significantly increased from 1 hour to 96 hours (4 days). This allows more time for retrieval of object values by ticker. The refactoring also activates the att_price_yahoo function for investment types that do not have an id of 16.