Refine stock selection with quantity check
The code is updated to improve the selection algorithm of stocks. Previously, stocks were selected if they existed in 'user_stocks' without any consideration to the quantity. Now, stocks are only considered if the quantity is more than 0 in 'user_stocks' database. This ensures only meaningful stocks are returned from the queries.