Add comments to improve avg_days calculation in investments
The comments were added to facilitate future adjustment of the calculation in the investment service. They detail how to reconsider the days of each purchase for a more accurate general measurement since the beginning of the investment. Suggestions include taking into account both the date of the movement and the calculation date. They also suggest considering the price for the correct computation of weights, including sold assets, and using the list of purchase movements instead of consolidated assets.